+++ src/modules/caregiver-info/caregiver-info.module.ts
@@ -2,12 +2,18 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { CaregiverInfo } from 'src/common/entities/caregiver.profile.entity';
+import { User } from 'src/common/entities/user.entity';
 
+import { ProfileModule } from '../profile/profile.module';
+import { UserService } from '../users/user.service';
+
+import { CaregiverInfoController } from './caregiver-info.controller';
 import { CaregiverInfoService } from './caregiver-info.service';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([CaregiverInfo])],
-  providers: [CaregiverInfoService],
-  exports: [CaregiverInfoService],
+  imports: [TypeOrmModule.forFeature([User, CaregiverInfo]), ProfileModule],
+  controllers: [CaregiverInfoController],
+  providers: [CaregiverInfoService, UserService],
+  exports: [CaregiverInfoService, UserService],
 })
 export class CaregiverInfoModule {}
