+++ src/modules/caregiver-info/types/detailed-caregiver-info.type.ts
@@ -1,0 +1,17 @@
+import { Certificate } from 'src/common/entities/certificate.entity';
+import { WorkExperience } from 'src/common/entities/work-experience.entity';
+import { Qualification } from 'src/common/enums/qualification.enum';
+
+export type DetailedCaregiverInfo = {
+  id: string;
+  firstName: string;
+  lastName: string;
+  isOpenToSeekerHomeLiving: boolean;
+  hourlyRate: number;
+  numberOfAppointments: number;
+  description: string;
+  videoLink: string;
+  services: Qualification[];
+  certificates: Certificate[];
+  workExperiences: WorkExperience[];
+};
