+++ src/modules/caregiver-info/types/parsed-params.type.ts
@@ -1,0 +1,11 @@
+export type ParsedParams = {
+  isOpenToSeekerHomeLiving: boolean;
+  isShowAvailableCaregivers: boolean;
+  country: string;
+  city: string;
+  address: string;
+  state: string;
+  zipCode: string;
+  utcOffset: string;
+  services: string[];
+};
